refactor: 유틸리티 메서드는 객체 생성을 막는 것이 좋음
